organize into modules
